feat: Separate HTML to Image logic into custom hook
- Extract HTML to image conversion logic into useConvertToImage custom hook
- Move image conversion and state management logic inside hook
- Reduce MainFrame component complexity and improve code reusability
