+++ src/components/File/SelectedFile/index.tsx
@@ -1,15 +1,11 @@
-import { useEffect, useRef, useState, createContext } from "react";
 import {
-  fileSysyemStore,
   PrChangedFileInfo,
   PrChangedFileStatusInfo,
 } from "@/stores/github.store";
-import { drawBoardStore } from "@/stores/drawBoard.store";
 import { editor } from "monaco-editor";
 import { Monaco } from "@monaco-editor/react";
 import { DrawBoard } from "@/components/Draw/DrawBoard";
 import { InitFile } from "@/components/File/SelectedFile/InitFile";
-import { toPng } from "html-to-image";
 import { BlockNote } from "@/components/BlockNote";
 import { RemovedFile } from "./RemovedFile";
 import { RenamedFile } from "./RenamedFile";
@@ -29,19 +25,6 @@ interface SelectedFileViewerProps extends PrChangedFileStatusInfo {
   ) => void;
 }
 
-interface BoardContextType {
-  convertToImage: () => void;
-  size: {
-    width: number;
-    height: number;
-  };
-}
-
-export const BoardContext = createContext<BoardContextType>({
-  convertToImage: () => {},
-  size: { width: 0, height: 0 },
-});
-
 const SelectedFileViewer = ({
   status,
   selectedCommitFile,
@@ -49,64 +32,6 @@ const SelectedFileViewer = ({
   onEditorMount,
   onSplitEditorMount,
 }: SelectedFileViewerProps) => {
-  // DrawBoard 컴포넌트를 항상 마운트된 상태로 유지하되 숨김 처리
-  const setIsImageAddedInfo = drawBoardStore(
-    (state) => state.setIsAddedImageInfo,
-  );
-  const elementRef = useRef<HTMLDivElement>(null);
-  const [size, setSize] = useState({ width: 0, height: 0 });
-  const setSelectedCommitFile = fileSysyemStore(
-    (state) => state.setSelectedCommitFile,
-  );
-
-  useEffect(() => {
-    const element = elementRef.current;
-    if (!element) return;
-
-    const resizeObserver = new ResizeObserver((entries) => {
-      for (const entry of entries) {
-        const { width, height } = entry.contentRect;
-        setSize({ width, height });
-      }
-    });
-
-    resizeObserver.observe(element);
-
-    return () => {
-      resizeObserver.disconnect();
-    };
-  }, [selectedCommitFile]);
-
-  const contextValue = {
-    convertToImage: () => {
-      if (!elementRef.current) return;
-      toPng(elementRef.current, { cacheBust: false })
-        .then((dataUrl: string) => {
-          sessionStorage.setItem("image", dataUrl);
-          setIsImageAddedInfo({
-            added: true,
-            size: {
-              width: size.width,
-              height: size.height,
-            },
-          });
-          setSelectedCommitFile({
-            additions: 0,
-            afterContent: "",
-            beforeContent: "",
-            deletions: 0,
-            filename: "MainDrawBoard",
-            language: "",
-            status: "init",
-          });
-        })
-        .catch((err: unknown) => {
-          console.error(err);
-        });
-    },
-    size,
-  };
-
   const fileComponent: {
     [key in PrChangedFileStatusInfo["status"]]: React.ReactNode;
   } = {
@@ -136,16 +61,7 @@ const SelectedFileViewer = ({
     return <BlockNote />;
   }
 
-  return (
-    <BoardContext.Provider value={contextValue}>
-      <div
-        ref={elementRef}
-        className="flex h-full w-full flex-col items-end justify-center"
-      >
-        {fileComponent[status] || <div>invalid</div>}
-      </div>
-    </BoardContext.Provider>
-  );
+  return <> {fileComponent[status] || <div>invalid</div>}</>;
 };
 
 export default SelectedFileViewer;
